apps/backend/services/orders/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Data Validation & Settings
pydantic
pydantic-settings

# Database (ORM & Driver)
sqlalchemy[asyncio]
asyncpg
alembic

# Messaging
kafka-python

# Utilities
python-dotenv
passlib[bcrypt] # For potential future password hashing if auth is co-located
python-jose[cryptography] # For JWT handling
